updated workshops links and added test file to test workshop links
