+++ lmfdb/templates/workshops.html
@@ -27,8 +27,7 @@ Sponsored by <i>Computational Representation Theory in Number Theory</i>, NSF gr
 
 <br>
 
-<li><a href="http://indico.ictp.it/event/a13219/material/1/0.pdf">Advanced School and 
-Workshop on L-functions and Modular Forms</a>,  <br>
+<li>Advanced School and Workshop on L-functions and Modular Forms</a>,  <br>
 September 1-12, 2014, <a href="https://www.ictp.it/">International Centre for Theoretical 
 Physics</a>, Trieste, Italy
 <br>
